Boilerplate pages
File structure for the new pages is ready. I spent more time learning and playing around with how to use Frank's ValidInput and ValidForm component, but there's still a lot of learning to be done.

I think I have some idea of how I wish to delegate pages to the rest of us as well as get my own hands dirty with some of the implementation. I'm feeling smarter already!